pyproject.toml
[project]
name = "cli-weight"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "plotext>=5.3.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "sqlmodel>=0.0.37",
]


[project.scripts]
weightcli = "main:main"
